Crate hartex_rust

A safe and lightweight real-time Kernel written in Rust. The Kernel is developed for cortex-m3/m4 based microcontrollers. The goal of the project was to develop a memory efficient, safe and lightweight real-time Kernel. Rust-lang was choosen due to its powerful compile-time checks. All the subsystems have been developed and manually tested. Language features like Traits, Generics helped define powerful and safe types. The Kernel uses statically allocated data structures to keep itself simple and fast. But using feature flags, the developer can enable dynamic memory allocation (for end-application and not Kernel itself). Cargo feature flags are used to configure constants such as maximum tasks, resources, etc.

Features Flags :

alloc : Enables use of dynamic data-structures.

tasks_8 : Max tasks in Kernel is set to 8.

tasks_16 : Max tasks in Kernel is set to 16.

tasks_32 : Max tasks in Kernel is set to 32.

resources_16 : Max resources in Kernel is set to 16.

resources_32 : Max resources in Kernel is set to 32.

resources_64 : Max resources in Kernel is set to 64.

messages_16 : Max messages in Kernel is set to 16.

messages_32 : Max messages in Kernel is set to 32.

messages_64 : Max messages in Kernel is set to 64.

events_16 : Max events in Kernel is set to 16.

events_32 : Max events in Kernel is set to 32.

events_64 : Max events in Kernel is set to 64.

Modules

config
Kernel configuration. Private

events
Kernel routines which assist in Event management.

kernel
Kernel module and routines declaration. Private

macros
Macro Definitions

messages
Kernel routines which assist in Inter-task Communication.

resources
Kernel routines which assist in Resource management.

semaphores
Kernel routines which assist in Inter-task Synchronization.

system
Kernel Data-structures. Private

tasks
Kernel routines which assist in Task management.

time
Kernel routines which assist in Time management.

types
Exports types defined across other Kernel modules.

util
Helper functions.

utils
Utility functions. Private

Macros

priv_execute
priv_execute! executes the code block only if the current context is in privileged mode.

spawn
The tasks must be looping infinitely and call task_exit whenever a particular task is done. This makes it complicated to create tasks and also might introduce undefined behavior if task_exit is not called. The spawn macro makes it easier to define tasks. It also defines a static variable of type TaskId, which corresponds to the task created.