Struct hartex_rust::system::event_manager::EventManager

pub struct EventManager {
    event_table: [Option<Event>; 32],
    curr: usize,
    ms_event_table: EventIndexTable,
    sec_event_table: EventIndexTable,
    min_event_table: EventIndexTable,
    hr_event_table: EventIndexTable,
    onoff_event_table: EventIndexTable,
}

Holds and Implements all Event management and dispatch methods.

Fields
event_table: [Option<Event>; 32]

This array holds the Event descriptors of all events

curr: usize

Points to the current empty slot in the event_table.

ms_event_table: EventIndexTable

An instance of EventIndexTable which holds list of EventIds of type EventTableType::MilliSec.

sec_event_table: EventIndexTable

An instance of EventIndexTable which holds list of EventIds of type EventTableType::Sec.

min_event_table: EventIndexTable

An instance of EventIndexTable which holds list of EventIds of type EventTableType::Min.

hr_event_table: EventIndexTable

An instance of EventIndexTable which holds list of EventIds of type EventTableType::Hour.

onoff_event_table: EventIndexTable

An instance of EventIndexTable which holds list of EventIds of type EventTableType::OnOff.

Methods

impl EventManager

pub const fn new() -> Self

Returns new instance of EventManager

pub fn sweep(&mut self, event_type: EventTableType)

This function dispatches all events mentioned in the EventIndexTable corresponding to the EventTableType.

pub fn execute_event(&mut self, event_id: EventId)

Takes the EventId and executes the corresponding event handler.

fn execute_opcode(&mut self, event_id: EventId)

This method executes all the operations corresponding to the opcode of the task.

pub fn enable_event(&mut self, event_id: EventId)

Enables an Event.

pub fn disable_event(&mut self, event_id: EventId)

Disables an Event.

pub fn create(
&mut self,
is_enabled: bool,
event_type: EventType,
threshold: u8,
event_counter_type: EventTableType
) -> Result<EventId, KernelError>

Creates a new event.

pub fn set_semaphore(
&mut self,
event_id: EventId,
sem: SemaphoreId,
tasks_mask: BooleanVector
) -> Result<(), KernelError>

Updates the opcode by setting the bit corresponding to Semaphore and stores the SemaphoreId and Tasks Boolean Vector.

pub fn set_tasks(
&mut self,
event_id: EventId,
tasks_mask: BooleanVector
) -> Result<(), KernelError>

Updates the opcode by setting the bit corresponding to Release Tasks and stores the tasks_mask Boolean Vector.

pub fn set_message(
&mut self,
event_id: EventId,
msg_id: MessageId
) -> Result<(), KernelError>

Updates the opcode by setting the bit corresponding to the Next event and stores the EventId.

pub fn set_next_event(
&mut self,
event_id: EventId,
next: EventId
) -> Result<(), KernelError>

Updates the opcode by setting the bit corresponding to Message and stores the MessageId.